#include<stdio.h> #include<math.h> int main()
{

double  p,r,t,ci;
scanf("%lf%lf%lf",&p,&r,&t);
ci=p*pow((1+(r/100)),t)-p;
printf("%.2lf",ci);

}

+++ Compound_Interest.c
@@ -4,9 +4,9 @@ int main()
 
 {
 
-float p,r,t,ci;
-scanf("%f%f%f",&p,&r,&t);
+double  p,r,t,ci;
+scanf("%lf%lf%lf",&p,&r,&t);
 ci=p*pow((1+(r/100)),t)-p;
-printf("%.2f",ci);
+printf("%.2lf",ci);
 
 }
